refactor: consolidate Slab1-4 views into a single Slab.vue
