--- src/views/Slab4.vue
@@ -1,11 +1,0 @@
-<script setup lang="ts">
-import Sidebar from '@/components/TeacherSidebar.vue'; 
-import PCLayout from '@/components/SLab4_Layout.vue';   
-</script>
-
-<template>
-  <div class="flex h-screen overflow-hidden">
-    <Sidebar class="flex-shrink-0" />
-    <PCLayout class="flex-1" />
-  </div>
-</template>
